# docs/backend_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Delay Prediction API",
    "description": "Готовые прогнозы задержек, активный транспорт и аналитика сегментов. Расчёты выполняют фоновые воркеры; HTTP backend читает результаты. В historical-режиме доступны состояние и запуск воспроизведения.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/predictions/{tr_id}/latest": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Получить последний прогноз одного ТС",
        "description": "Вернуть последний прогноз ТС с оценкой свежести ready или stale.\n\nОтсутствие прогноза: HTTP 404; недоступность БД: HTTP 503.",
        "operationId": "get_latest_for_vehicle_predictions__tr_id__latest_get",
        "parameters": [
          {
            "name": "tr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tr Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}